+++ visual-js/visual-nightwatch/index.js
@@ -1,8 +1,0 @@
-const chalk = require('chalk');
-const { DiffStatus } = require('@saucelabs/visual');
-
-console.log(
-  `\n${chalk.yellow('Loaded @saucelabs/nightwatch-sauce-visual-service')}\n`,
-);
-
-module.exports.DiffStatus = DiffStatus;