+++ visual-python/src/saucelabs_visual/typing.py
@@ -1,7 +1,8 @@
-from dataclasses import dataclass, InitVar
+from dataclasses import dataclass, InitVar, asdict
 from enum import Enum
 from typing import List, Union, Literal
 
+from selenium.webdriver.remote.webelement import WebElement
 from typing_extensions import TypedDict, NotRequired
 
 
@@ -62,25 +63,43 @@ class IgnoreRegion(IgnoreBase):
     name: Union[str, None] = None
 
 
-class FullPageConfig(TypedDict):
-    delayAfterScrollMs: NotRequired[int]
+@dataclass
+class IgnoreElementRegion(IgnoreBase):
+    element: Union[WebElement, List[WebElement]]
+    diffingOptions: Union[DiffingOptions, None] = None
+    name: Union[str, None] = None
+
+    def as_dict_array(self):
+        element_array = self.element if isinstance(self.element, List) else [self.element]
+        return [
+            {
+                "id": element.id,
+                "diffingOptions": self.diffingOptions,
+                "name": self.name,
+            } for element in element_array
+        ]
+
+
+@dataclass
+class FullPageConfig:
+    delayAfterScrollMs: Union[int, None] = None
     """
     Delay in ms after scrolling and before taking screenshots. A slight delay can be helpful for
     websites leveraging lazy loading.
     """
-    hideAfterFirstScroll: NotRequired[List[str]]
+    hideAfterFirstScroll: Union[List[str], None] = None
     """
-    Adjust address bar padding on iOS and Android for viewport cutout.
+    A list of CSS selectors for elements to hide after the first scroll.
     """
-    disableCSSAnimation: NotRequired[bool]
+    disableCSSAnimation: Union[bool, None] = None
     """
     Disable CSS animations and the input caret for the site under test.
     """
-    hideScrollBars: NotRequired[bool]
+    hideScrollBars: Union[bool, None] = None
     """
     Hide all scrollbars for the site under test.
     """
-    scrollLimit: NotRequired[int]
+    scrollLimit: Union[int, None] = None
     """
     Limit the number of screenshots taken for scrolling and stitching. Default and max value is 10.
     """